requirements.txt
alt_profanity_check==1.3.2
diffusers==0.25.1
nltk==3.8.1
numpy==1.24.3
pandas==2.0.2
Pillow==9.5.0
Pillow==10.3.0
plotly==5.18.0
PyYAML==6.0
PyYAML==6.0.1
Requests==2.32.2
streamlit==1.28.1
textblob==0.17.1
torch==2.0.1+cu118
transformers==4.38.0.dev0
